# GymZZC/swagger.yaml
openapi: 3.0.0
info:
  title: API de Gestión de Entrenamiento
  version: 1.0.0
  description: API para gestionar usuarios, rutinas, ejercicios y progreso de entrenamiento
servers:
  - url: https://api.tuentrenador.com/v1
    description: Servidor de producción
  - url: http://localhost:3000/api
    description: Servidor de desarrollo
paths:
...
  /usuarios:
    get:
      summary: Obtiene usuarios con filtros opcionales
      description: Retorna lista de usuarios con posibilidad de filtrar por nombre, rol o email
      parameters:
        - in: query
          name: nombre
          schema:
            type: string
          description: Filtro por nombre (búsqueda parcial)
        - in: query
          name: rol
          schema:
            type: string
            enum: [cliente, entrenador, admin]
          description: Filtro por rol de usuario
        - in: query
          name: email
          schema:
            type: string
          description: Búsqueda exacta por email
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'
    post:
      summary: Crea un nuevo usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioInput'
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
  /usuarios/{id}:
    get:
      summary: Obtiene un usuario por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del usuario
      responses:
        '200':
          description: Detalles del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
    put:
      summary: Actualiza un usuario existente
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioInput'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
    delete:
      summary: Elimina un usuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Usuario eliminado

  # =============================================
  # ENDPOINTS DE RUTINAS
  # =============================================
  /rutinas:
    get:
      summary: Obtiene todas las rutinas
      responses:
        '200':
          description: Lista de rutinas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rutina'
    post:
      summary: Crea una nueva rutina
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RutinaInput'
      responses:
        '201':
          description: Rutina creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rutina'
...
  /rutinas/{rutina_id}/ejercicios:
    get:
      summary: Obtiene ejercicios de una rutina
      parameters:
        - in: path
          name: rutina_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de ejercicios en la rutina
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RutinaEjercicio'
    post:
      summary: Añade un ejercicio a una rutina
      parameters:
        - in: path
          name: rutina_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RutinaEjercicioInput'
      responses:
        '201':
          description: Ejercicio añadido a la rutina
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RutinaEjercicio'
  /rutinas/{rutina_id}/ejercicios/{ejercicio_id}:
    get:
      summary: Obtiene detalles de un ejercicio en una rutina
      parameters:
        - in: path
          name: rutina_id
          required: true
          schema:
            type: integer
        - in: path
          name: ejercicio_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalles del ejercicio en la rutina
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RutinaEjercicio'
    put:
      summary: Modifica un ejercicio en una rutina
      parameters:
        - in: path
          name: rutina_id
          required: true
          schema:
            type: integer
        - in: path
          name: ejercicio_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RutinaEjercicioInput'
      responses:
        '200':
          description: Ejercicio actualizado en la rutina
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RutinaEjercicio'
    delete:
      summary: Quita un ejercicio de una rutina
      parameters:
        - in: path
          name: rutina_id
          required: true
          schema:
            type: integer
        - in: path
          name: ejercicio_id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Ejercicio eliminado de la rutina

  # =============================================
  # ENDPOINTS DE PROGRESO
  # =============================================
...
  /progreso/usuario/{usuario_id}:
    get:
      summary: Obtiene todo el progreso de un usuario
      parameters:
        - in: path
          name: usuario_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de progresos del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Progreso'
...
components:
  schemas:
...
    Usuario:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: Ana López
        email:
          type: string
          example: ana@example.com
        rol:
          type: string
          enum: [cliente, entrenador, admin]
          example: cliente
        fecha_registro:
          type: string
          format: date
          example: '2023-05-15'
      required: [id, nombre, email, rol]
    UsuarioInput:
      type: object
      properties:
        nombre:
          type: string
          example: Carlos Ruiz
        email:
          type: string
          example: carlos@example.com
        rol:
          type: string
          enum: [cliente, entrenador, admin]
          example: entrenador
      required: [nombre, email, rol]

    # ======== RUTINA ========
    Rutina:
      type: object
      properties:
        id:
          type: integer
          example: 5
        usuario_id:
          type: integer
          example: 1
        entrenador_id:
          type: integer
          example: 2
        descripcion:
          type: string
          example: Rutina de fuerza inicial
      required: [id, usuario_id, entrenador_id, descripcion]
    RutinaInput:
      type: object
      properties:
        usuario_id:
          type: integer
          example: 1
        entrenador_id:
          type: integer
          example: 2
        descripcion:
          type: string
          example: Rutina de definición
      required: [usuario_id, entrenador_id, descripcion]

    # ======== EJERCICIO ========
...
    RutinaEjercicio:
      type: object
      properties:
        rutina_id:
          type: integer
          example: 5
        ejercicio_id:
          type: integer
          example: 8
        series:
          type: integer
          example: 4
        repeticiones:
          type: integer
          example: 12
        peso_ejercicio:
          type: integer
          example: 20
      required: [rutina_id, ejercicio_id, series, repeticiones]
    RutinaEjercicioInput:
      type: object
      properties:
        ejercicio_id:
          type: integer
          example: 8
        series:
          type: integer
          example: 4
        repeticiones:
          type: integer
          example: 12
        peso_ejercicio:
          type: integer
          example: 20
      required: [ejercicio_id, series, repeticiones]

    # ======== PROGRESO ========
    Progreso:
      type: object
      properties:
        id:
          type: integer
          example: 25
        usuario_id:
          type: integer
          example: 1
        rutina_id:
          type: integer
          example: 5
        ejercicio_id:
          type: integer
          example: 8
        fecha:
          type: string
          format: date
          example: '2023-06-10'
        series_completadas:
          type: integer
          example: 4
        repeticiones_completadas:
          type: integer
          example: 10
        peso_utilizado:
          type: number
          format: float
          example: 22.5
        notas:
          type: string
          example: Última serie con ayuda
      required: [id, usuario_id, rutina_id, ejercicio_id, fecha, series_completadas, repeticiones_completadas]
...